Sling export of a hierarchy of pages -- introduction of the v3/page
