+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/internal/models/v3/HierarchyComponentContextWrapper.java
@@ -13,7 +13,7 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-package com.adobe.cq.wcm.core.components.internal.models.v2;
+package com.adobe.cq.wcm.core.components.internal.models.v3;
 
 import java.util.Set;
