+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/internal/models/v2/PageImplTest.java
@@ -20,20 +20,14 @@ import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
-import org.apache.sling.api.SlingHttpServletRequest;
-import org.apache.sling.api.resource.Resource;
-import org.apache.sling.models.factory.ModelFactory;
-import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.osgi.framework.Version;
 import org.powermock.reflect.Whitebox;
 
-import com.adobe.cq.export.json.hierarchy.HierarchyNodeExporter;
 import com.adobe.cq.wcm.core.components.Utils;
 import com.adobe.cq.wcm.core.components.models.NavigationItem;
 import com.adobe.cq.wcm.core.components.models.Page;
@@ -47,8 +41,6 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.when;
 
 public class PageImplTest extends com.adobe.cq.wcm.core.components.internal.models.v1.PageImplTest {
@@ -56,16 +48,13 @@ public class PageImplTest extends com.adobe.cq.wcm.core.components.internal.mode
     private static final String TEST_BASE = "/page/v2";
     private static final String REDIRECT_PAGE = ROOT + "/redirect-page";
 
-    private static final String PAGE_CHILD = PAGE + "/child";
-
     private static ClientLibrary mockClientLibrary;
 
     private static MockProductInfoProvider mockProductInfoProvider = new MockProductInfoProvider();
 
-    @BeforeClass
-    public static void setUp() {
+    protected static void internalSetUpV2(String testBase) {
+        internalSetUp(CONTEXT, testBase, ROOT);
 
-        internalSetUp(CONTEXT, TEST_BASE, ROOT);
         mockClientLibrary = Mockito.mock(ClientLibrary.class);
 
         when(mockClientLibrary.getPath()).thenReturn("/apps/wcm/core/page/clientlibs/favicon");
@@ -74,59 +63,15 @@ public class PageImplTest extends com.adobe.cq.wcm.core.components.internal.mode
         CONTEXT.registerInjectActivateService(mockProductInfoProvider);
     }
 
+    @BeforeClass
+    public static void setUp() {
+        internalSetUpV2(TEST_BASE);
+    }
+
     @Test
     public void testPage() throws ParseException {
         Page page = getPageUnderTest(PAGE);
-        Calendar calendar = Calendar.getInstance();
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
-        calendar.setTime(sdf.parse("2016-01-20T10:33:36.000+0100"));
-        assertEquals(page.getLastModifiedDate().getTime(), calendar.getTime());
-        assertEquals("en-GB", page.getLanguage());
-        assertEquals("Templated Page", page.getTitle());
-        assertEquals(DESIGN_PATH, page.getDesignPath());
-        assertNull(page.getStaticDesignPath());
-        String[] keywordsArray = page.getKeywords();
-        assertEquals(3, keywordsArray.length);
-        Set<String> keywords = new HashSet<>(keywordsArray.length);
-        keywords.addAll(Arrays.asList(keywordsArray));
-        assertTrue(keywords.contains("one") && keywords.contains("two") && keywords.contains("three"));
-        assertArrayEquals(new String[] {"coretest.product-page", "coretest.product-page-js-head"}, page.getClientLibCategories());
-        assertArrayEquals(new String[] {"coretest.product-page-js-head"}, page.getClientLibCategoriesJsHead());
-        assertArrayEquals(new String[] {"coretest.product-page"}, page.getClientLibCategoriesJsBody());
-        assertEquals("product-page", page.getTemplateName());
-        Utils.testJSONExport(page, Utils.getTestExporterJSONPath(TEST_BASE, PAGE));
-    }
-
-    @Test
-    public void testChildren() {
-        Page page = getPageUnderTest(PAGE);
-        ModelFactory modelFactory = Mockito.mock(ModelFactory.class);
-        Whitebox.setInternalState(page, "modelFactory", modelFactory);
-
-        Page childPage = getPageUnderTest(PAGE_CHILD);
-        when(modelFactory.getModelFromWrappedRequest(any(SlingHttpServletRequest.class), any(Resource.class), eq(Page.class))).thenReturn(childPage);
-
-        Map<String, ? extends HierarchyNodeExporter> children = page.getExportedChildren();
-        Assert.assertEquals(2, children.size());
-        Assert.assertEquals(childPage, children.get(PAGE_CHILD));
-
-    }
-    @Test
-    public void testGetRootUrl() {
-        // no parent
-        Page rootPage = getPageUnderTest(PAGE);
-        Assert.assertEquals(CONTEXT_PATH + PAGE + ".model.json", rootPage.getRootUrl());
-
-        // with parent
-        Page childPage = getPageUnderTest(PAGE_CHILD);
-        Assert.assertEquals(CONTEXT_PATH + PAGE + ".model.json", childPage.getRootUrl());
-    }
-
-    @Test
-    public void testGetRootModel() {
-        // no parent
-        Page page = getPageUnderTest(PAGE);
-        Assert.assertEquals(page, page.getRootModel());
+        assertPage(page);
     }
 
     @Test(expected = UnsupportedOperationException.class)
@@ -177,4 +122,31 @@ public class PageImplTest extends com.adobe.cq.wcm.core.components.internal.mode
     private Page getPageUnderTest(String pagePath) {
         return super.getPageUnderTest(Page.class, pagePath);
     }
+
+    protected String getTestBase() {
+        return TEST_BASE;
+    }
+
+    protected void assertPage(Page page) throws ParseException {
+        Calendar calendar = Calendar.getInstance();
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
+        calendar.setTime(sdf.parse("2016-01-20T10:33:36.000+0100"));
+        assertEquals(page.getLastModifiedDate().getTime(), calendar.getTime());
+        assertEquals("en-GB", page.getLanguage());
+        assertEquals("Templated Page", page.getTitle());
+        assertEquals(DESIGN_PATH, page.getDesignPath());
+        assertNull(page.getStaticDesignPath());
+        String[] keywordsArray = page.getKeywords();
+        assertEquals(3, keywordsArray.length);
+        Set<String> keywords = new HashSet<>(keywordsArray.length);
+        keywords.addAll(Arrays.asList(keywordsArray));
+        assertTrue(keywords.contains("one") && keywords.contains("two") && keywords.contains("three"));
+        assertArrayEquals(new String[] {"coretest.product-page", "coretest.product-page-js-head"}, page.getClientLibCategories());
+        assertArrayEquals(new String[] {"coretest.product-page-js-head"}, page.getClientLibCategoriesJsHead());
+        assertArrayEquals(new String[] {"coretest.product-page"}, page.getClientLibCategoriesJsBody());
+        assertEquals("product-page", page.getTemplateName());
+        Utils.testJSONExport(page, Utils.getTestExporterJSONPath(getTestBase(), PAGE));
+    }
+
+
 }
