+++ content/src/content/jcr_root/apps/core/wcm/components/page/v3/page/head.resources.html
@@ -1,0 +1,26 @@
+<!--/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  ~ Copyright 2017 Adobe Systems Incorporated
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
+<template data-sly-template.favicons="${ @ path ='The path to the client library resources folder'}">
+    <link rel="icon" sizes="192x192" href="${path}/icon-192x192.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="${path}/apple-touch-icon-180x180.png">
+    <link rel="apple-touch-icon" sizes="167x167" href="${path}/apple-touch-icon-167x167.png">
+    <link rel="apple-touch-icon" sizes="152x152" href="${path}/apple-touch-icon-152x152.png">
+    <link rel="apple-touch-icon" sizes="120x120" href="${path}/apple-touch-icon-120x120.png">
+    <meta name="msapplication-square310x310logo" content="${path}/icon-310x310.png">
+    <meta name="msapplication-wide310x150logo" content="${path}/icon-310x150.png">
+    <meta name="msapplication-square150x150logo" content="${path}/icon-150x150.png">
+    <meta name="msapplication-square70x70logo" content="${path}/icon-70x70.png">
+</template>
